add,make

+++ src/main/java/com/project/yjshop/domain/token/RefreshTokenRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Integer> {
     Optional<RefreshToken> findByToken(String token);
 }
